pyproject.toml
[project]
name = "gen-au"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.5.5",
    "faiss-cpu==1.8.0",
    "fastapi>=0.135.1",
    "langchain==0.2.14",
    "langchain-anthropic>=0.1.23",
    "langchain-community==0.2.12",
    "langchain-core>=0.2.43",
    "langchain-google-genai==1.0.8",
    "langchain-huggingface>=0.0.3",
    "langchain-mistralai==0.1.12",
    "langchain-openai==0.1.22",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "requests>=2.32.5",
    "sentence-transformers==3.0.1",
    "tiktoken==0.7.0",
    "transformers>=4.57.6",
    "uvicorn>=0.41.0",
]
